pyproject.toml
[tool.poetry]
name = "meshbbs"
version = "0.0.1"
description = "A meshtastic BBS system"
readme = "README.md"
authors = ["Josh Bressers <josh@bres.net>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
pypubsub = "^4.0.3"
meshtastic = "^2.5.0"
peewee = "^3.17.6"

[tool.poetry.scripts]
meshbbs = "meshbbs.server:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
